Updated working version of the DungeonMate Web app
A tool for Dungeon Masters (DM) to help accompany their DnD (Dungeons & Dragons) campaign. It helps them build an immersive ambiance during the game.

Gif logo showing on the Homepage. When clicked, the user gets taken to a page with the map being shown. The map can be scroll zoomed or pinched and panned. There are audio file buttons beside it that can be played and paused. These sounds can be stacked on top of each other, depending on the scene that the DM decides for the players.

Still in development.